astro/requirements.txt
astronomer-cosmos>=1.0.2
dbt-snowflake